[REFACTOR] features 레이어 filter-movie 수정
